services/nutrition-ai-service/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# OpenAI
openai>=1.40.0
httpx==0.24.1

# Authentication & Security
python-jose[cryptography]==3.3.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Rate Limiting
slowapi==0.1.9

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
